Created header and implementation file from test_joystick_driver.cpp
