+++ igvc_utils/src/mocking_utils/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(mocking_utils mock_subscriber.cpp)
